feat: complete cuda mode, and improve user input customization

+++ alignment/constants.hpp
@@ -2,7 +2,7 @@
 
 #include <cinttypes>
 
-constexpr int16_t MATCH = 2;
-constexpr int16_t MISMATCH = -1;
-constexpr int16_t GAP_OPEN = -2;
-constexpr int16_t GAP_EXTEND = -1;
+constexpr int8_t MATCH = 2;
+constexpr int8_t MISMATCH = -1;
+constexpr int8_t GAP_OPEN = -2;
+constexpr int8_t GAP_EXTEND = -1;
